+++ packages/core/src/map.ts
@@ -12,30 +12,39 @@ export const map =
   <Error>(
     source: LazyPromise<Value, Error>,
   ): LazyPromise<NewValue, Error | NewError> =>
-    createLazyPromise((resolve, reject, fail) => {
-      let dispose: (() => void) | undefined;
-      const disposeOuter = source.subscribe(
-        (value) => {
-          let newValueOrPromise;
-          try {
-            newValueOrPromise = callback(value);
-          } catch (error) {
-            fail(error);
-            return;
-          }
-          if (isLazyPromise(newValueOrPromise)) {
-            dispose = newValueOrPromise.subscribe(resolve, reject, fail);
-          } else {
-            resolve(newValueOrPromise);
-          }
-        },
-        reject,
-        fail,
-      );
-      if (!dispose) {
-        dispose = disposeOuter;
-      }
-      return () => {
-        dispose!();
-      };
-    });
+    createLazyPromise(
+      (resolve: ((value: NewValue) => void) | undefined, reject, fail) => {
+        let dispose: (() => void) | undefined;
+        const disposeOuter = source.subscribe(
+          (value) => {
+            let newValueOrPromise;
+            try {
+              newValueOrPromise = callback(value);
+            } catch (error) {
+              if (!resolve) {
+                throw error;
+              }
+              fail(error);
+              return;
+            }
+            if (!resolve) {
+              return;
+            }
+            if (isLazyPromise(newValueOrPromise)) {
+              dispose = newValueOrPromise.subscribe(resolve, reject, fail);
+            } else {
+              resolve(newValueOrPromise);
+            }
+          },
+          reject,
+          fail,
+        );
+        if (!dispose) {
+          dispose = disposeOuter;
+        }
+        return () => {
+          resolve = undefined;
+          dispose!();
+        };
+      },
+    );